src/examples/simple-image-classification/requirements.txt
certifi==2020.6.20
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.20.2
olefile==0.46
Pillow==8.0.0
pip==21.1.3
setuptools==49.6.0.post20210108
six==1.16.0
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
typing-extensions==3.10.0.0
wheel==0.36.2
